moved logic of reading/writing state to file_io and state_manager
